snap/snaplock: move locking routines
The per-snap lock routines were defined in the mount interface package
long ago, when they were still used there. They are only referenced from
snap-update-ns now so it's sensible to move them.

I will need the same routine in overlord/snapstate so I picked
snap/snaplock (due to the dependency on osutil) as the appropriate
location.
