+++ snap/snaplock/lock.go
@@ -17,37 +17,30 @@
  *
  */
 
-package mount_test
+package snaplock
 
 import (
+	"fmt"
 	"os"
-	"strings"
-
-	. "gopkg.in/check.v1"
+	"path/filepath"
 
 	"github.com/snapcore/snapd/dirs"
-	"github.com/snapcore/snapd/interfaces/mount"
+	"github.com/snapcore/snapd/osutil"
 )
 
-type lockSuite struct{}
-
-var _ = Suite(&lockSuite{})
-
-func (s *lockSuite) SetUpTest(c *C) {
-	dirs.SetRootDir(c.MkDir())
+// lockFileName returns the name of the lock file for the given snap.
+func lockFileName(snapName string) string {
+	return filepath.Join(dirs.SnapRunLockDir, fmt.Sprintf("%s.lock", snapName))
 }
 
-func (s *lockSuite) TearDownTest(c *C) {
-	dirs.SetRootDir("")
-}
-
-func (s *lockSuite) TestOpenLock(c *C) {
-	lock, err := mount.OpenLock("name")
-	c.Assert(err, IsNil)
-	defer lock.Close()
-
-	_, err = os.Stat(lock.Path())
-	c.Assert(err, IsNil)
-
-	c.Check(strings.HasPrefix(lock.Path(), dirs.SnapRunLockDir), Equals, true, Commentf("wrong prefix for %q, want %q", lock.Path(), dirs.SnapRunLockDir))
+// OpenLock creates and opens a lock file associated with a particular snap.
+func OpenLock(snapName string) (*osutil.FileLock, error) {
+	if err := os.MkdirAll(dirs.SnapRunLockDir, 0700); err != nil {
+		return nil, fmt.Errorf("cannot create lock directory: %s", err)
+	}
+	flock, err := osutil.NewFileLock(lockFileName(snapName))
+	if err != nil {
+		return nil, err
+	}
+	return flock, nil
 }
